Refactor FG_RPM and RS485 modules to improve buffer management and data handling; update RPM reference value calculations and remove RS485_enum.h

+++ FG_RPM/FG_RPM.h
@@ -24,7 +24,7 @@
 
 typedef struct {
   tmr_type *tmr_list[MAX_TIMERS];
-  uint8_t timer_count;  // 使用者設定的 timer 數量
+  uint8_t timer_count;
 
   uint32_t exint_line;
   uint8_t motor_phase;
@@ -32,14 +32,13 @@ typedef struct {
   uint32_t tmr_clk;
   uint16_t tmr_div[MAX_TIMERS];
   uint16_t tmr_period[MAX_TIMERS];
-  float rpm_ref_val[MAX_TIMERS];
+  int32_t rpm_ref_val[MAX_TIMERS];
   uint8_t rpm_ref_val_index;
 
   bool first_count_flag;
   uint16_t first_count[MAX_TIMERS];
   uint16_t sample;
   uint8_t interrupt_count[MAX_TIMERS];
-
 } FgParam_t;
 
 void FgInit(FgParam_t *fg_param);
